pyproject.toml
[project]
name = "cva-mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.4",
    "openai>=1.0.0",
    "pyyaml>=6.0.2",
]
